requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
pymongo[srv]
fastapi-mail
email-validator
pandas
motor
apscheduler
uuid
csv
